requirements.txt
anthropic>=0.49.0
beautifulsoup4>=4.12.0
click>=8.1.0
python-dotenv>=1.0.0
requests>=2.31.0
rich>=13.7.0
